Add custom handling for room shares to the list of sharees
Like done for other types of shares, room shares are now explicitly
described as such in the UI.

The avatar used is the image provided in the "shareWithAvatar" property
of the share. If none is given then the avatar is the first letter of
the display name of the room share with a coloured background seeded
from the room token. If the display name of the room is empty then no
letter is shown in the avatar; no special handling is done in that case.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>