Extend the auth page for public shares to request the password by Talk
When the authentication page for a public share is shown and the sharer
enabled the option "send password by Talk" now the sharee can request
the password to the sharer by Nextcloud Talk directly from the
authentication page.

The public share authentication controller just checks if the option is
set and Talk is enabled and, in that case, it appends a Talk template to
the current one. That template, which is part of the Talk app, uses
JavaScript in the browser to inject the Talk UI as needed in the page
generated by the server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>